content-security-policy/to-port-from-blink/connect-src-websocket-blocked-expected.txt

CONSOLE ERROR: Refused to connect to 'ws://localhost:8880/echo' because it violates the following Content Security Policy directive: "connect-src ws://127.0.0.1:8880".

Pass
